Fix GradioWorkbook Height
# Fix GradioWorkbook Height

Gradio components are rendered within an iframe in HF spaces. With the editorBackground minHeight currently set to 100vh, this makes the containing div's height expand outside the bounds of the iframe in gradio and result in an endless page.

To fix, just make the height 100% of the containing element (iframe in this case).

## Testing:
In styles.css in gradio-workbook, I set
```
div.editorBackground {
    min-height: 400px;
    height: 100%;
    background-color: green
}
```
to override the `editorBackground` styles with higher specificity. Then, publish the component to load in the workspace:

https://github.com/lastmile-ai/aiconfig/assets/5060851/f436bb87-a0d8-4887-95c0-68839237feb3


Note that we can now see the bottom of the page, as well as the notifications! Note that notifications are relative to the bottom right of the root element (iframe in this case), so only visible if the bottom of the iframe is in view. I don't think we can really fix that in this scenario.