Set Model Selector Label to 'HuggingFace Task' for 'gradio' Mode
# Set Model Selector Label to 'HuggingFace Task' for 'gradio' Mode

Per dogfood feedback, it's confusing for the model selector label and placeholder to say 'Model' while the parser is registered based on task name. As a temporary improvement, this PR uses 'HuggingFace Task" terminology instead of model. Long-term, once the model -> model parser mapping UX is available, we should undo this and have the model copy back in the selector.

## Updated for 'gradio' Mode:
<img width="230" alt="Screenshot 2024-01-26 at 4 35 27 PM" src="https://github.com/lastmile-ai/aiconfig/assets/5060851/ab6ee0d1-bf46-4fae-a098-e79f1487ef1b">

## Local / non-gradio is unchanged:
<img width="1304" alt="Screenshot 2024-01-26 at 4 35 09 PM" src="https://github.com/lastmile-ai/aiconfig/assets/5060851/9cbc21a9-9ddf-4ac8-96cc-effe1deae353">