Hardcode default models for HuggingFace Remote Inference schemas
HuggingFace Remote Inference Model parsers utilized the Remote Inference endpoint.

When a model is not provided, it uses a default model. For now, hardcode the default model into the prompt schema, such that the user sees the "default model". This will help improve user experience

Found the models by calling the Inference Endpoint API:
<img width="383" alt="Screenshot 2024-01-26 at 4 16 10 PM" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/82c5badf-d611-4e00-b258-e64cf4557c12">

There probably is a way to do something similar in Typescript and such implement it in the schema.


## Testplan:

Verify selecting the 7 tasks populates the model field
https://github.com/lastmile-ai/aiconfig/assets/141073967/69ed36f5-2259-41b9-a535-d0819bb91997