Dynamically Set Settings/Parameters Height to Match Prompt Input/Output Height
# Dynamically Set Settings/Parameters Height to Match Prompt Input/Output Height

This PR implements dynamic max height for the model settings and parameters contents to be bounded by the height of the prompt input/output cell container. I wasn't able to get anything working with pure css so this implementation requires some javascript -- not ideal, due to potential perf hit, but it works.

The main idea is that:
1. In PromptContainer, use a ref for the element containing the input/outputs and calculate its height whenever input or outputs change (using debounce so we don't do this constantly). Set this height in an exported const map, mapped under the prompt id
2. In the PromptActionBar, set the max height of the contents to one of:
a) the calculated height minus height of surrounding action bar contents (84px, found with some trial/error logging)
b) 300px if a) is less than 300px since we want the minimum height to be reasonable enough to show sufficient settings

## Testing:
Min height of 300px for empty new prompt
<img width="1375" alt="Screenshot 2024-01-28 at 3 46 35 PM" src="https://github.com/lastmile-ai/aiconfig/assets/5060851/ac42a05a-fc3d-430d-bf8e-86775396aa8c">


Dynamic resizing:

https://github.com/lastmile-ai/aiconfig/assets/5060851/591fe137-edfb-4c7b-9b6d-be6d282efb15