[8/n] Add init command to vscode extension
The way we package dependencies for the vscode extension is by using pip to install the `python-aiconfig` package, as well as any aiconfig extensions we want to package together by default.

This diff introduces the `check_requirements.py` script which knows how to parse a requirements.txt file and check if the dependencies and versions listed in it are already installed on the system.

The `init` command then does the following:
* Checks if python is installed (if not, it prompts the user to install).
* Checks if pip is installed on the system (this should always be installed if python has been installed, but is done as a sanity check)
* Checks if given the requirements.txt, any packages need to be installed or updated.
* If an update is required, it installs the dependencies using pip.

This is displayed to the user in the form of a vscode progress notification.

Finally, we create an AIConfig output channel that displays logs in the Output pane. This will be used by the aiconfig server as well to display detailed logs.

Test Plan:
* See video in comment.
* Also did some manual stress testing of `check_requirements.py` script -- made sure it handles ==, <=, >=, <, > operators, weird spaces, and comments in the requirements.txt file.