[4/n] Add 'start' command to aiconfig server and CLI
The current aiconfig CLI has an 'edit' command which starts both the frontend and backend servers, and also loads the aiconfig from disk.

In the same vein, this diff introduces a 'start' command that is used to just start the aiconfig server. This is used by the vscode extension to spawn a server when an aiconfig file is opened.

Unlike the 'edit' command, 'start':
* Doesn't load the aiconfig on server init. Instead, the /api/load_content endpoint is called to initialize.
* Doesn't search for open ports. The vscode client is responsible for finding an open port. This is because the client needs to be in control of the server lifecycle.

A small change that was also added to this diff: using `dotenv` to load an env file.

Test Plan:
* At the top of stack