[gradio] Move Styles into GradioTheme
# [gradio] Move Styles into GradioTheme

Ok, this is a long stack of incremental changes to hopefully get gradio styles to look decent / match intended GradioTheme styles once and for all.

This PR moves the styles from gradio's styles.css into the theme so that we can hopefully remove the styles.css in gradio. Note that we nest them under the `div.editorBackground` (meaning, they apply to any elements with those classes and which are also descendants of the div with editorBackground class). This should ensure a high specificity on their selectors.


Check that the buttons have the correct styles:
![Screenshot 2024-01-31 at 11 30 32 PM](https://github.com/lastmile-ai/aiconfig/assets/5060851/c30725bf-044a-4c85-9b59-8142e1f6c6ff)
![Screenshot 2024-01-31 at 11 30 57 PM](https://github.com/lastmile-ai/aiconfig/assets/5060851/b65b0935-dd06-48a5-a4e1-e61786466fa9)