[editor][ez] Fix AddPromptButton Width
# [editor][ez] Fix AddPromptButton Width

Fix the AddPromptButton so that it no longer extends beyond the left of the prompts. Requires both adding left margin (to offset its left side to match start of the prompt cells) and removing the 100% width (to prevent the right side from extending outside of the container right padding). The width is still the full width of the prompts without explicitly setting width

## Before
![Screenshot 2024-02-02 at 5 59 23 PM](https://github.com/lastmile-ai/aiconfig/assets/5060851/1112b68c-f779-417c-93c5-0b391dcef306)

## After
![Screenshot 2024-02-02 at 5 57 53 PM](https://github.com/lastmile-ai/aiconfig/assets/5060851/07879a14-fecd-4f10-82dd-655876f0b904)