[editor][1/4] Set Up Notification Provider/Context
# [editor][notifications][1/4] Set Up Notification Provider/Context

For the vs code extension, we want to leverage the vs code notification UX instead of mantine for showing notifications. To support that in a usable way, this PR stack sets up a Notification context which will propagate provided `showNotification` function to any components that request it from context. If there is no `showNotification` provided to the `AIConfigEditor` component we will default to the mantine `showNotification` function as usual.

This PR sets up the context, managing which `showNotification` function to call, and providing it through the context provider.

## Testing:
- No-op here. Make sure notifications still work as usual