[python] temporary Fix Gemini Event loop
Gemini api under the hood saves a reference to the async event loop. This messes with the local editor because we create a new event loop on api/run.

This diff reinitializes the "client" at the run step for the Gemini model parser. I put client in quotes because its not called a client but its analogous to the openai client abstraction.

## Testplan

Run a gemini prompt twice inside local editor.
Before:
<img width="1598" alt="Screenshot 2024-02-05 at 5 36 26 PM" src="https://github.com/lastmile-ai/gradio-workbook/assets/141073967/5ed779f4-61b2-4595-a34c-11992a2c055c">

After:
run twice, no event loop error
https://github.com/lastmile-ai/gradio-workbook/assets/141073967/39d1d4b2-90c5-4ee1-b1a4-bc35d8528d60