[python] Pin Huggingface-hub dependency version
From #1156, not all versions of the python package huggingface-hub is compatible with the python sdk for AIConfig. We utilize the `InferenceClient` which was introduced sometime in 2023.

This diff pins it on > 0.20.3.

Current use of huggingface-hub is within the hugging face text generation model parser (core model parser).

## Testplan

I ran pip list on my computer to check which version I am developing with locally.
```
ankush@ap-mbp aiconfig % pip list | grep huggingface-hub
huggingface-hub                               0.20.3
```