rfc: first serialize assistant message as prompt
A user was building with AIConfig and was attempting to serialize a set of prompts with OpenAI.

Their messages array looked like this
```
{ "role": "assistant", "content": "Welcome, ask me anything." },
{ "role": "user", "content": "Say this is a test" },
...
```

Notice how the first message was an assistant message! Previously serializing this with config.serialize() was not supported.

This diff adds minimal support by allowing serialization of the first assistant message, and creates a prompt with an empty input, and the assistant message as the output.

## testplan
Execute the user's test code. Also tested deserialize() and run() and they work as expected. No functional changes on that side

| <img width="546" alt="Screenshot 2024-02-07 at 11 42 31 AM" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/06299485-a5ae-4a9c-adf6-e6f21293d8ed">|<img width="456" alt="Screenshot 2024-02-07 at 11 41 53 AM" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/e60c1cc6-cf7a-4e6a-8e20-e6ad6db866bf">|