[rfc] Prevent 'Error starting aiconfig server' notification when closing w/ unsaved changes
# [rfc] Prevent 'Error starting aiconfig server' notification when closing w/ unsaved changes

Currently, whenever an aiconfig editor webview is closed with unsaved changes, a notifcation shows at the bottom right:
![Screenshot 2024-02-08 at 6 29 06 PM](https://github.com/lastmile-ai/aiconfig/assets/5060851/9e45fced-a60e-42e9-8a2f-e8f65cfcd27f)

This is caused by the following sequence of events:
1. changeDocumentSubscription happens with the "unsaved" changes being undone; this results in updating webview & server
2. server /load request is sent as a result
3. Another changeDocumentSubscription happens with no changed content (can probably just skip -- in all cases?)
4. Webview disposed, kills editor process
5. /load Request (STILL IN FLIGHT) fails, causing the notification

I spent some time but wasn't able to figure out why changeDocumentSubscription was happening (note -- NOT happening from the notifying dirty). Seems like this could potentially be a KP - https://github.com/microsoft/vscode/issues/81902 but not concrete.

For now, RFC on this, we could just handle the server request error specially in this case when the webview is already disposed (and ignore it).


https://github.com/lastmile-ai/aiconfig/assets/5060851/4e986654-31d9-4192-bba7-d07fe0756448

Make sure the error still shows for legit cases (e.g. opening config without correct parsers installed):

https://github.com/lastmile-ai/aiconfig/assets/5060851/912b8123-137f-4e8e-ab64-b05438adb44f