Clear model settings (and use prompt input defaults) when model_name is changed
Ok this is extremely hacky becuse we are CLEARING the existing model settings so and relying on the client input schema default to take over. If the client default and server default are misaligned, then this could lead to more issues, so we need to be careful. This is also just cumbersome to maintain, but I feel it's the fastest immediate short-term solution for this, even if it's annoying

## Test Plan
Make sure to run `yarn && yarn build` from the editor client file, then go to `aiconfig` and run this:
```
aiconfig_path=./cookbooks/Gradio/huggingface.aiconfig.json
parsers_path=./cookbooks/Gradio/aiconfig_model_registry.py
aiconfig edit --aiconfig-path=$aiconfig_path --server-port=8080 --server-mode=debug_servers --parsers-module-path=$parsers_path
```
For now, I only set default for HuggingFaceText2ImageDiffusor (because I can feel myself being tired and not thinking properly), but this is a proof of concept that it works, just need to set it for all the Remote inference HF input schemas


https://github.com/lastmile-ai/aiconfig/assets/151060367/357054b3-7425-4b96-9c88-364280a0fc20