[gradio] Fix Spinner Colors
# [gradio] Fix Spinner Colors

The loading spinner on the run button, download and share buttons is not clearly visible in gradio right now, especially in light mode:

https://github.com/lastmile-ai/aiconfig/assets/5060851/a7204248-1a8a-4f3a-8d12-e553dd66c03d


To fix, we need to override the stroke for the svg that's used. Also, for running prompt button, updated to make the square in the runPromptButton fill with the gradio color instead of white.

Test in light & dark themeMode:

https://github.com/lastmile-ai/aiconfig/assets/5060851/489745c0-f5bb-4f17-8cfd-d817adc78c9b


https://github.com/lastmile-ai/aiconfig/assets/5060851/e37e543c-c43d-43f6-bf22-4bce4ea08644