Activate extension after VSCode startup
This will trigger the python dependencies to be installed, saving time for when we first open an AIConfig file. This will call the `activate()` method in the `extension.ts` file, which contains this line:

```
// Run the setup command on activation
  vscode.commands.executeCommand(COMMANDS.INIT);
```

## Test Plan
No functional changes, but activation now gets triggered right away instead of opening an AIConfig file. Notice that opening the file is now faster and doesn't need to wait for activate steps to take effect

Before


After