Activate extension after VSCode startup
This will trigger the python dependencies to be installed, saving time for when we first open an AIConfig file. This will call the `activate()` method in the `extension.ts` file, which contains this line:

```
// Run the setup command on activation
  vscode.commands.executeCommand(COMMANDS.INIT);
```

## Test Plan
No functional changes, but activation now gets triggered right away instead of opening an AIConfig file. This still takes a bit of time to startup because we need to run the flask server, but if I didn't have dependencies installed, it would be huge time savings!

Before

https://github.com/lastmile-ai/aiconfig/assets/151060367/fb45a79a-7593-4745-bfa2-41fa0856470c

After

https://github.com/lastmile-ai/aiconfig/assets/151060367/c164d180-c7ce-4062-bbe9-5452449e0ffe