[vscode][2/n] Setup Env Variables: Define env path
Opening a window for user to enter in an `.env` path. Using the home dir for now, but can change to workspace if you want, not a big deal.

Next diffs will implement a static file template and use that when env path doesn't exist. Diff after that I'll deal with existing env path situation

## Test Plan

https://github.com/lastmile-ai/aiconfig/assets/151060367/cceadc3a-cd20-4164-b8b2-c8d11b36b5bb