3/n Update Initialize to invoke select interpreter
This diff updates the initialize flow to include the Python interpreter selection step. It ensures dependencies are installed only after the interpreter selection is finalized, accommodating the no-op change where the selected interpreter remains unchanged.

Note: If the extension is not activated, invoking the Initialize AIConfig command for an unknown reason installs dependencies on the pre-selected interpreter path


## testplan

Extension initializes on Opening An AIConfig

Test select a new interpreter

Test reselecting the same interpreter

https://github.com/lastmile-ai/aiconfig/assets/141073967/1a51ff89-3889-4fd6-8964-c341453ddf68