[vscode] Remove Initialization Flow from Extension Activation
This diff removes the Initialization flow from ocurring on Extension Activation.

The landing page directs the user to invoke this command.

The Opening an AIConfig may invoke this if dependencies are not installed (not implemented yet)

Invoking the Initialize Command will no longer automatically install dependencies on the pre-selection python interpreter.

## Testplan

https://github.com/lastmile-ai/aiconfig/assets/141073967/04031cdc-2c9d-47e8-b43d-a9d2f6aa769d