Create listener to open info dialog for when python interpreter path changes
Haven't added the actual implementation for restarting server. Ryan is working on that now so will connect to it next PR when it's ready

Some additional context that Ankush and I discussed:

If user selects an invalid interpreter, that's fine because:
1. They AIconfig flow will be invalid and fail (upon restarting server), but we have checks to pester them to update their interpreter to the correct one via the `installDependencies()` flow (defined in https://github.com/lastmile-ai/aiconfig/pull/1307)

## Test Plan

https://github.com/lastmile-ai/aiconfig/assets/151060367/b9219c15-5843-4bf7-888f-357b20faff6a

Also tested this does not trigger the button to be shown when selecting an interpreter gets triggered because we haven't selected it before. This is good, because it would otherwise be annoying to refresh the window when it's already "active". It's technically not active yet because we haven't yet done the python checks and started the server

https://github.com/lastmile-ai/aiconfig/assets/151060367/5be17f39-8d7c-470f-b04e-25093187a8f6