[local editor] Don't poll server status when in read only view
In order to prevent showing status error message when in read only view, this diff makes sure that the aiconfig editor client does not poll the server status when in read only mode.

See #1320 for more context

## Testplan
1. Manually set readonly prop to True
2. Load local editor
3. kill the python process that is hosting the aiconfig server.

| Before  | After |
| ------------- | ------------- |
| <img width="1913" alt="after" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/ce3487ba-0174-4ca7-b58e-7e10aed6409d">  | <img width="1909" alt="before" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/cba7b817-d09e-435f-b2e8-da7c13109459"> |

Note: the `killpython` command in my testplan is a custom alias I have defined that tries to kill all pid's that have the descriptor "python" in its name.