[1/n] Implement Editor Server as Class
# [1/n] Implement Editor Server as Class

Currently we maintain EditorServer state as an object for the extension, accessing the process and server url from the object. In subsequent PR, we will need to expose the ability to restart the editor server. Instead of creating and exposing a restart function on the object, let's just move the editor server into a class so that we have a proper instance with methods to leverage for starting, stopping and (next PR) restarting.

## Testing:
Open an aiconfig, ensure the server starts and I can run a prompt
Close the config, verify in logging that the server process is properly killed
Open config again. See server start up on 8000 (i.e. the port is open again after server was killed)