Fix error starting server on new aiconfig yaml
When you create a new aiconfig yaml, the server fails to start.

This is because of changes to the flow where we are calling `updateServerState` in the launch flow of the editor, which in turn calls `getModeForDocument`. This function uses the file extension to determine if the document is JSON or YAML, but in the case of Untitled documents, there is no file URI.

Since it is reasonable to expect this flow for Untitled documents, the fix here is to use the document's content to determine the mode (try parsing it as JSON or as YAML).

https://github.com/lastmile-ai/lastmile/assets/25641935/1c96d099-d970-4e5f-9b58-3768451e591c