[RFC] Handle creating files as *.aiconfig.yaml or *.aiconfig.json files
Untitled files have a lot of constraints on how they can be set up. Currently it defaults to saving the file as *.json or *.yaml.

This RFC is about saving the file as *.aiconfig.json or *.aiconfig.yaml.

Unfortunately this isn't perfect either, because it now saves the file as untitled-1.aiconfig.json (where `1` gets incremented). No file save as dialog is shown in the flow.

An alternative is to have the `Create New` command ask for a file name or path on creation (and enforce the file ending in the accepted extensions)