Force walkthrough to open when extension is activated the first time
This is kind of a hack. I'm not sure why the extension isn't opening when it first installs, but whatever this will ensure it runs. Adding a global config for `vscode-aiconfig.version`, and also changed the activate settings to activate upon vscode startup. This is fine because we moved the initialize step outside of here so there's no negative consequences of activating it.

## Test Plan
Ok I don't know why I can't test directly bvy running the activation command, but instead I just hooked this up to the 'Set API Keys" command and ran it lol