[vscode] Prompt User to restart Vscode after installing python
As part of the vscode extension installation flow, if the user doesn't have a valid python interpreter, they are prompted with a vscode notification asking them if they would like to install python.

If they do install python, vscode will need to be restarted in order to have the new interpreter available to select.

This diff shows the user a notification with the option to restart vscode for them, if they click on `Install Python` as mentioned above.

RFC: Python Extension API exposes  functionality to [Detect new environments discovered in the system](https://github.com/microsoft/vscode-python/wiki/Python-Environment-APIs#detect-new-environments-or-changes-to-the-interpreters-discovered-in-the-system) but I felt it would be easier to just have the user restart vscode.

## Testplan:

https://github.com/lastmile-ai/aiconfig/assets/141073967/6568d9ca-a7d1-47dd-abf4-8a2276658061