Run the install dependencies command if we're restarting the server
Before we wouldn't do this, so we could change the interpreter, it could trigger user to retsrat the server, but that wouldn't allow them to get in a non-restart state because the dependencies weren't installed. This only happens on first time when a new interpreter is selected AFTER we open the AIconfig file (since we check for running dependencies on file open). After you've selected interpreter, it's saved to cache and when you open the file, it'll auto-install the second time. This PR changes it to work on the first time!

## Test Plan

Before

https://github.com/lastmile-ai/aiconfig/assets/151060367/5db776b7-ce65-4c1a-a39e-69fc4c144d99


After

https://github.com/lastmile-ai/aiconfig/assets/151060367/285e215e-0095-4157-bb4e-dff29c8b346d